requirements.txt
whisper==1.1.10
moviepy==1.0.3
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
Pillow==10.0.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
flask==2.3.3
gunicorn==21.2.0
werkzeug==2.3.7 